Check for existence of custom template

+++ src/lambda/write-template.js
@@ -1,6 +1,9 @@
-let { readFileSync, writeFile } = require('fs')
+let { existsSync, readFileSync, writeFile } = require('fs')
 
 module.exports = function writeTemplate ({ handlerFile, template }, callback) {
+  if (!existsSync(template)) {
+    throw Error(`Custom function template not found: ${template}`)
+  }
   let body = readFileSync(template).toString()
   writeFile(handlerFile, body, callback)
 }
